SOV-746: Fix email notification settings dialog (#2495)
* Trigger email notification settings dialog from anywhere

* Fix subscriptions updating + cleanup
